Refactored response conversion to a separate class
