+++ lib/data_services_api.rb
@@ -10,6 +10,7 @@ require 'data_services_api/query_generator'
 require 'data_services_api/value'
 require 'data_services_api/service_exception'
 require 'data_services_api/sapint_converter.rb'
+require 'data_services_api/dsapi_response_converter.rb'
 
 # :nodoc:
 module DataServicesApi
